+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/ChapterImpl.kt
@@ -31,6 +31,8 @@ class ChapterImpl : Chapter {
 
     override var last_page_read: Int = 0
 
+    override var scroll_position: Int = 0
+
     override var pages_left: Int = 0
 
     override var date_fetch: Long = 0